Reimplemented hashmap using a generic macro.
This commit updates the hashmap implementation which was only able to save
the following combinations of key/value:
 * string / void *
 * uint64_t / void *

Using the macro found here https://github.com/neovim/neovim/blob/master/src/nvim/map.c (44b2cef)
allows us to use more generic hash values and keys. From this, it follows that it is
easier to safe dispatch information in the hashtable, and from that follows that the
memory management is simplified.