Update message deserialize request and response implementation
The deserializer function alloced memory for the response, this
makes it hard to free and keep track of the memory. now the caller
of the deserializer functions is responible for creating a
request or response object. This change makes memory management
for requests and responses more easy. Some tests needed update
to work with the new implementation.